+++ lib/oli/resources.ex
@@ -121,9 +121,8 @@ defmodule Oli.Resources do
   end
 
   # returns a list of resource ids that refer to activity references in a page
-  def activity_references(%Revision{content: %{"model" => model}} = _page) do
-    model
-    |> Enum.filter(&(&1["type"] == "activity-reference"))
+  def activity_references(%Revision{content: content} = _page) do
+    Oli.Resources.PageContent.flat_filter(content, &(&1["type"] == "activity-reference"))
     |> Enum.map(& &1["activity_id"])
   end