+++ CHANGELOG.md
@@ -7,11 +7,11 @@
 - Add support for detecting problematic database queries
 - Allow adaptive pages to render without application chrome
 - Save cookie preferences in delivered courses
+- Add support for page content grouping
 
 ### Bug fixes
 
 - Fix help and logo links on register institution page, change help form to modal
-
 - Add missing database indexes, rework resolver queries
 
 ## 0.9.0 (2021-4-22)